requirements.txt
cohere
openai
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
youtube-transcript-api